feat: added routes for monthly and yearly
